removal of redundant code and files
