+++ src/assets/index.js
@@ -1,37 +1,20 @@
 import logo from "./logo1.png";
-import backend from "./backend.png";
-import creator from "./creator.png";
-import mobile from "./mobile.png";
-import web from "./web.png";
 import github from "./github.png";
 import menu from "./menu.svg";
 import close from "./close.svg";
-
 import css from "./tech/css.png";
 import docker from "./tech/docker.png";
-import figma from "./tech/figma.png";
 import git from "./tech/git.png";
 import html from "./tech/html.png";
 import javascript from "./tech/javascript.png";
 import mongodb from "./tech/mongodb.png";
 import nodejs from "./tech/nodejs.png";
 import reactjs from "./tech/reactjs.png";
-import redux from "./tech/redux.png";
 import tailwind from "./tech/tailwind.png";
-import typescript from "./tech/typescript.png";
-import threejs from "./tech/threejs.svg";
-
-import meta from "./company/meta.png";
-import shopify from "./company/shopify.png";
-import starbucks from "./company/starbucks.png";
 import sce from "./company/sce.png";
 import sjsu from "./company/sjsu.png";
 import hornet from "./company/hornet.png";
-import tesla from "./company/tesla.png";
 import discord from "./tech/discord.png";
-import carrent from "./carrent.png";
-import jobit from "./jobit.png";
-import tripguide from "./tripguide.png";
 import game from "./tech/game.png";
 import bot from "./bot.png";
 import swanky from "./swanky.png";
@@ -42,34 +25,19 @@ import transit from "./transit.png";
 
 export {
   logo,
-  backend,
   game,
-  creator,
-  mobile,
-  web,
   github,
   menu,
   close,
   css,
   docker,
-  figma,
   git,
   html,
   javascript,
   mongodb,
   nodejs,
   reactjs,
-  redux,
   tailwind,
-  typescript,
-  threejs,
-  meta,
-  shopify,
-  starbucks,
-  tesla,
-  carrent,
-  jobit,
-  tripguide,
   sce,
   sjsu,
   hornet,
