+++ src/components/About.jsx
@@ -1,38 +1,8 @@
 import React from "react";
-import { Tilt } from "react-tilt";
 import { motion } from "framer-motion";
 import { SectionWrapper } from '../hoc';
 import { styles } from "../style";
-import { services } from "../constants";
 import { fadeIn, textVariant } from "../utils/motion";
-
-const ServiceCard = ({ index, title, icon }) => (
-  <Tilt className='xs:w-[230px] w-full'>
-    <motion.div
-      variants={fadeIn("right", "spring", index * 0.5, 1.5)}
-      className='w-full green-pink-gradient p-[1px] rounded-[20px] shadow-card'
-    >
-      <div
-        options={{
-          max: 45,
-          scale: 1,
-          speed: 450,
-        }}
-        className='bg-tertiary rounded-[20px] py-5 px-12 min-h-[230px] flex justify-evenly items-center flex-col'
-      >
-        <img
-          src={icon}
-          alt='web-development'
-          className='w-16 h-16 object-contain'
-        />
-
-        <h3 className='text-white text-[20px] font-bold text-center'>
-          {title}
-        </h3>
-      </div>
-    </motion.div>
-  </Tilt>
-);
 
 const About = () => {
   return (
@@ -68,12 +38,6 @@ const About = () => {
             <img src="/matt.png" alt="Matt Estrada" className="w-full max-w-[360px] rounded-full shadow-lg" />
           </div>
         </motion.div>
-      </div>
-
-      <div className='mt-20 flex flex-wrap justify-center gap-10'>
-        {services.map((service, index) => (
-          <ServiceCard key={service.title} index={index} {...service} />
-        ))}
       </div>
     </>
   );
