+++ src/components/Experience.jsx
@@ -1,9 +1,7 @@
 import React from "react";
 import { VerticalTimeline, VerticalTimelineElement, } from "react-vertical-timeline-component";
 import { motion } from "framer-motion";
-
 import "react-vertical-timeline-component/style.min.css";
-
 import { styles } from "../style";
 import { experiences } from "../constants";
 import { SectionWrapper } from "../hoc";
