--- src/components/Loader.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Loader = () => {
-  return (
-    <div>Loader</div>
-  )
-}
-
-export default Loader
