+++ src/components/Projects.jsx
@@ -1,9 +1,6 @@
 import React from "react";
-import { Tilt } from "react-tilt";
 import { motion } from "framer-motion";
-
 import { styles } from "../style";
-import { github } from "../assets";
 import { SectionWrapper } from "../hoc";
 import { projects } from "../constants";
 import { fadeIn, textVariant } from "../utils/motion";
