+++ src/hoc/SectionWrapper.jsx
@@ -1,5 +1,4 @@
 import { motion } from "framer-motion";
-
 import { styles } from "../style";
 import { staggerContainer } from "../utils/motion";
