+++ src/hoc/index.js
@@ -1,3 +1,2 @@
 import SectionWrapper from "./SectionWrapper";
-
 export { SectionWrapper };
